CurrencyLayer로부터 받는 환율 정보를 String으로 받아서 BigDecimal로 처리 후 response

+++ src/main/java/com/seogineer/wirebarleyjointest/dto/ExchangeRateResponse.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class ExchangeRateResponse {
-    private Double USDKRW;
-    private Double USDJPY;
-    private Double USDPHP;
+    private Float USDKRW;
+    private Float USDJPY;
+    private Float USDPHP;
 }
